Rename index to app Change denon.json start cmd to new name Refactor app
