+++ src/server.ts
@@ -1,0 +1,2 @@
+import app from './app.ts'
+
